HBASE-19388 - Incorrect value is being set for Compaction Pressure in RegionLoadStats object inside HRegion class

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/HRegion.java
@@ -7189,8 +7189,8 @@ public class HRegion implements HeapSize, PropagatingConfigurationObserver, Regi
         stats.setHeapOccupancy((int)(occupancy * 100));
       }
     }
-    stats.setCompactionPressure((int)rsServices.getCompactionPressure()*100 > 100 ? 100 :
-                (int)rsServices.getCompactionPressure()*100);
+    stats.setCompactionPressure((int) (rsServices.getCompactionPressure() * 100 > 100 ? 100
+        : rsServices.getCompactionPressure() * 100));
     return stats.build();
   }
